requirements.txt
homeassistant>=2024.1.0
aiohttp>=3.8.0
requests>=2.31.0
pyyaml>=6.0.1 